civic_salon/pyproject.toml
[tool.poetry]
name = "civic-salon"
version = "0.1.0"
description = ""
authors = ["Yuway Chen <yuway_chen@trendmicro.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
pydantic = "^2.9.2"
databricks-vectorsearch = "0.40"
flashrank = "^0.2.9"
openai = "^1.51.0"
langchain-community = "^0.3.1"
langchain-core = "^0.3.9"
tiktoken = "^0.8.0"
langgraph = "^0.2.34"
langchain-openai = "^0.2.2"
langchain-databricks = "^0.1.0"
uvicorn = "^0.31.1"
beautifulsoup4 = "^4.12.3"
selenium = "^4.25.0"
webdriver-manager = "^4.0.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
